Major refactoring of Tags, Templates, CompiledDocument
Tags: Implemented the rendering of html and tex when the tag's contents are a tag themselves. This generated a bug with process_paragraphs. Added tests to the core tags.
Tags: Add newlines for tex of heading tags. Updated tests.
Tags: Added text wrapping to paragraphs in tex mode for a cleaner output.
Tags: Changed Tag behavior and added tests for the rendering of html and tex tags when the contents are for an invalid type. This now raises a TagError.
Tags: Added the notes tags
Templates: Implemented a Tufte latex template (tufte.tex)
AST: Added tests for process_ast with simple strings as input (strings without sub-tags)
AST: Added test for process_paragraphs for simple strings.
CompiledDocument: Moved a test (test_bad_pdflatex) to the convert submodule, as this is a more appropriate place for this test.