Refactor process_paragraph to be more conservative in defining paragraphs, fix the CLI options
process_paragraph: Refactored so that the contents of a parapgraph wouldn't simply be another tag. Paragrapgs are collections of text and possibly some tags.
eq: Updated template (eq.tex) and tests.
tag settings: add the 'pre' environment.
server: allow the port number and target_list to be specified from the CLI.