Simplify document target_list and remove caching, fixed render_compiled bug, fixed translate_path bug, refactored label references when a target is missing
Document: Disabled caching of targets and target_list since these are cheap to check from the context (and easily invalidated).
Document: Fixed a bug in render_compiled that didn't correctly copy the compiled target.
Document: Fixed bug in the translate_path utility function to return a path for an actual target of a document, rather than an existing patch. This bug would give the wrong paths for non-target files retrieved from the webserver, like media files.
Document: Added the project's title in the render_tree_html utility function.
Labels: Refactored the reference method to not add a link when there is no target available for the label's document.
Server: Fixed bugs in rendering documents on every refresh and added debugging info.