Added mtime checking for the Document and LabelManager
Document: rename sub_document to subdocument, added the mtime for the label_manager as a test for the render_required method.
LabelManager: Add tracking of the modification time when labels are added, removed or changed.
Macro: Removed the author macro, as it caused some recursion problems.
Tags: Fixed preamble Titlepage tag that inserted the incorrect Tex command for the titlepage.
CSS updates