Refactored label manager with modification times and access the chapter/section labels.
AST: Add flatten_ast utility function and tests.
Document: Convert context variables of documents to weakrefs so that a document object can be deleted when removed from the subdocuments. Added tests.
Document: Added a test to 'render_required' for the mtime of labels.
Document: Fixed bug in the rendering of documents. The document is written whenever a render is done to keep the src_filepath from being newer than the target if the target doesn't change.
LabelManager: Refactor so that the label manager collects labels and a new method, register_labels, makes them accessible to tags. The register_labels method also populates weakrefs to chapter and section labels for labels, making these available to label formatted strings.
Labels: Added the label_latest_mtime function for getting the newest (largest) modification time of labels in an AST. Added tests.
Utils: Added a function to easily calculate a hash for a string.
Tags: Updated caption and heading tags to work with the new label manager format.