Refactor mtimes for documents, tags and labels
ast utils: move flatten_ast to the tag utilities.
label: refactor the update of mtimes to use the document's mtime, by default, and to simplify all of the properties to __getattr__ and __setattr__ methods.
document: store the mtime in the context, instead as an attribute, so that it's accessible to tags and labels.
document: refactor to use new label mtime scheme.
tag: refactor the mtime method to return the latest mtime between the context (from the document) and the mtimes of the labels.