Refactor subdocument management and labels
Document: Changed the subdocuments attribute to an OrderedDict, and removed the _subdocuments_src_filepaths list, which was difficult to manage. Subdocument order is now saved in the subdocuments ordered dict directly.
Document: Updated the document to remove invokations to the register_labels.
Document: Added the documents_dict method, which works directly with the subdocuments ordered dict and the documents_list method.
LabelManager: Moved the register_labels method to the get_label and get_labels methods, rather than the document, to more easily deal with labels.
Document tests: Refactored the test files, and completed the render_required test.