Fix tox tests for py35 and py36, refactor html tag attribute order
Tox: Update setup.cfg and setup.py to properly install disseminate in virtualenvs.
Attributes: kwargs_attributes now returns an ordered dict instead of a dict to preserve the order of attributes. This is needed for reproducible testing.
Document: Updates the ensure that path objects are always strings.
Document.utils: Update render_tree_html to preserve the order of html attributes.
Tags and Caption: Update the html method to preserve the order of attributes in html tags.
Document Tree tests: Fix a possible test error when trying to figure out whether a new object was created. Previously the object's id was used, which could come up to the same id when an object is re-created.