Refactor the context to include tags
AST: Added process_context_tags to process context (header) entries that could be converted to a tag.
Document: Replaced the process_context_toc to the new process_context_tags function in the AST.
Tags and Labels: Renamed the Chapter heading to 'Branch.' This was done because disseminate could be used for projects that aren't books and don't have chapters. Branches are the highest Heading level.
Toc: Delete the process_context_toc function since it's no longer needed.