Major refactoring to place arbitrary tags in the context and refactor header parsing
Document: Move the AST to the 'body' attribute in the context.
Header: Refactor to remove dependency on pyyaml and allow more flexible headers.
Tag: Set the lxml module to return safe strings instead of normal strings.
Template: Added the 'keep_trailing_newline' option to the environment to make sure a newline is added at the end of a rendered file.