Add 'replace' option for contexts
context.Context: Add the 'replace' attribute, which is a set of context entries for mutables that should be replaced (overwritten) rather than appended.
document.Document: Remove the target_list property, which is redundant.
document.DocumentContext: Add targets, and inactive_tags to the list of validation types, and placed these in the set of 'replaced' items in the context.
settings: Converted the target and inactive_tags to sets from the default_context.
document.process.ProcessContextHeaders: update to work with the new targets listing in documents
tags.TagFactory: Update the TagFactory to use the new 'inactive_tags' entry in the context
tags.navigation: Update navigation tags to work with the removal of the targets_list from documents